+++ src/main/java/com/verdantartifice/primalmagick/datagen/items/ItemModelProviderPM.java
@@ -100,6 +100,10 @@ public class ItemModelProviderPM extends ModelProvider<ItemModelBuilderPM> {
     }
     
     protected void registerModels(HolderLookup.Provider lookupProvider) {
+        // Generate crop items
+        this.basicItem(ItemsPM.HYDROMELON_SEEDS.get());
+        this.basicItem(ItemsPM.HYDROMELON_SLICE.get());
+        
         // Generate salted food items
         this.itemWithParent(ItemsPM.SALTED_BAKED_POTATO.get(), Items.BAKED_POTATO);
         this.itemWithParent(ItemsPM.SALTED_COOKED_BEEF.get(), Items.COOKED_BEEF);