+++ src/main/java/com/verdantartifice/primalmagick/datagen/loot_tables/AbstractBlockLootTableProvider.java
@@ -2,8 +2,10 @@ package com.verdantartifice.primalmagick.datagen.loot_tables;
 
 import java.util.Collections;
 import java.util.HashSet;
+import java.util.OptionalInt;
 import java.util.Set;
 import java.util.function.BiConsumer;
+import java.util.function.Function;
 import java.util.stream.Collectors;
 
 import org.apache.logging.log4j.LogManager;
@@ -23,6 +25,7 @@ import net.minecraft.world.item.enchantment.Enchantments;
 import net.minecraft.world.level.block.Block;
 import net.minecraft.world.level.block.SlabBlock;
 import net.minecraft.world.level.block.state.properties.SlabType;
+import net.minecraft.world.level.storage.loot.IntRange;
 import net.minecraft.world.level.storage.loot.LootPool;
 import net.minecraft.world.level.storage.loot.LootTable;
 import net.minecraft.world.level.storage.loot.entries.LootItem;
@@ -30,6 +33,7 @@ import net.minecraft.world.level.storage.loot.entries.LootPoolEntryContainer;
 import net.minecraft.world.level.storage.loot.functions.ApplyBonusCount;
 import net.minecraft.world.level.storage.loot.functions.ApplyExplosionDecay;
 import net.minecraft.world.level.storage.loot.functions.CopyNbtFunction;
+import net.minecraft.world.level.storage.loot.functions.LimitCount;
 import net.minecraft.world.level.storage.loot.functions.SetItemCountFunction;
 import net.minecraft.world.level.storage.loot.predicates.BonusLevelTableCondition;
 import net.minecraft.world.level.storage.loot.predicates.ExplosionCondition;
@@ -37,6 +41,7 @@ import net.minecraft.world.level.storage.loot.predicates.LootItemBlockStatePrope
 import net.minecraft.world.level.storage.loot.predicates.LootItemCondition;
 import net.minecraft.world.level.storage.loot.providers.nbt.ContextNbtProvider;
 import net.minecraft.world.level.storage.loot.providers.number.ConstantValue;
+import net.minecraft.world.level.storage.loot.providers.number.NumberProvider;
 import net.minecraft.world.level.storage.loot.providers.number.UniformGenerator;
 import net.minecraftforge.common.ToolActions;
 import net.minecraftforge.common.loot.CanToolPerformAction;
@@ -77,6 +82,10 @@ public abstract class AbstractBlockLootTableProvider extends BlockLootSubProvide
     private void registerLootTableBuilder(Block block, LootTable.Builder builder) {
         this.registeredBlocks.add(ForgeRegistries.BLOCKS.getKey(block));
         this.add(block, builder);
+    }
+    
+    protected void registerLootTableBuilder(Block block, Function<Block, LootTable.Builder> builderGenerator) {
+        this.registerLootTableBuilder(block, builderGenerator.apply(block));
     }
     
     protected void registerEmptyTable(Block block) {
@@ -143,6 +152,16 @@ public abstract class AbstractBlockLootTableProvider extends BlockLootSubProvide
         LootPool.Builder pulseBuilder = LootPool.lootPool().setRolls(ConstantValue.exactly(1.0F)).add(LootItem.lootTableItem(ItemsPM.HEARTWOOD.get()))
                 .when(LootItemBlockStatePropertyCondition.hasBlockStateProperties(block).setProperties(StatePropertiesPredicate.Builder.properties().hasProperty(AbstractPhasingLogBlock.PULSING, true)));
         LootTable.Builder tableBuilder = LootTable.lootTable().withPool(logBuilder).withPool(pulseBuilder);
+        this.registerLootTableBuilder(block, tableBuilder);
+    }
+    
+    protected void registerSplittingTable(Block block, Item splitItem, NumberProvider splitCount, OptionalInt maxWithFortune) {
+        var builder = LootItem.lootTableItem(splitItem).apply(SetItemCountFunction.setCount(splitCount));
+        if (maxWithFortune.isPresent()) {
+            builder = builder.apply(ApplyBonusCount.addUniformBonusCount(Enchantments.BLOCK_FORTUNE)).apply(LimitCount.limitCount(IntRange.upperBound(maxWithFortune.getAsInt())));
+        }
+        builder = builder.apply(ApplyExplosionDecay.explosionDecay());
+        LootTable.Builder tableBuilder = LootTable.lootTable().withPool(LootPool.lootPool().setRolls(ConstantValue.exactly(1.0F)).add(LootItem.lootTableItem(block).when(HAS_SILK_TOUCH).otherwise(builder)));
         this.registerLootTableBuilder(block, tableBuilder);
     }